RTK/requirements.txt
#Essential versions list for the packages used here
future==0.18.2
numpy
pandas
matplotlib
# pygmaps-enhanced==0.2.0 #Normal pygmaps will not work as it has been descontinued
gmaps==0.9.0
folium==0.13.0
iso8601==1.1.0
pynmeagps==1.0.4
pyserial==3.5
PyYAML==6.0
pyspark==3.3.2
utm==0.7.0